% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{extract_jason}
\alias{extract_jason}
\title{extract_jason}
\usage{
extract_jason(yahoo_url)
}
\arguments{
\item{yahoo_url}{url with the data to be extracted}
}
\value{
a data.table with the data extracted
}
\description{
Extract the data from the jason file
}
\keyword{internal}
