% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_symbols}
\alias{get_symbols}
\title{get_symbols}
\usage{
get_symbols(symbols, from, to, frequency)
}
\arguments{
\item{symbols}{character vector with the symbols to be queried}

\item{from}{character date from which the data will be queried}

\item{to}{character date to which the data will be queried}

\item{frequency}{character value with the frequency to be used in the query}
}
\value{
a data.table with the data extracted
}
\description{
Get the data from yahoo finance
}
\details{
Parameter frequency: Admitted values are 'daily', 'weekly', 'monthly', 'hourly'
}
\examples{
\dontrun{get_symbols(c("AAPL", "MSFT"), "2019-01-01", "2019-12-31", "daily")}
}
