% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_yahoo_url}
\alias{get_yahoo_url}
\title{get_yahoo_url}
\usage{
get_yahoo_url(symbol, from, to, interval)
}
\arguments{
\item{symbol}{character value with the symbol to be queried}

\item{from}{date from which the data will be queried}

\item{to}{date to which the data will be queried}

\item{interval}{time interval to be used in the query}
}
\description{
Build the url to query yahoo finance
}
\keyword{internal}
