% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{time_interval}
\alias{time_interval}
\title{time_interval}
\usage{
time_interval(interval)
}
\arguments{
\item{interval}{Admitted values are 'daily', 'weekly', 'monthly', 'hourly'}
}
\description{
Select the time interval to be used in the query
}
\keyword{internal}
