% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_intervals}
\alias{time_intervals}
\title{Time Series Frequencies}
\format{
\subsection{\code{time_intervals}}{

A character vector with 25 elements:
\describe{
\item{daily}{Daily frequency, "1d"}
\item{weekly}{Weekly frequency, "1wk"}
\item{monthly}{Monthly frequency, "1mo"}
\item{hourly}{Hourly frequency, "1h"}
\item{1minute, 1minutes, \verb{1 minute}, \verb{1 minutes}, \verb{1 min}, \verb{1 min}}{One minute frequency, "1m"}
\item{2minutes, \verb{2 minutes}, \verb{2 min}}{Two minutes frequency, "2m"}
\item{5minutes, \verb{5 minutes}, \verb{5 min}}{Five minutes frequency, "5m"}
\item{15minutes, \verb{15 minutes}, \verb{15 min}}{Fifteen minutes frequency, "15m"}
\item{30minutes, \verb{30 minutes}, \verb{30 min}}{Thirty minutes frequency, "30m"}
\item{60minutes, \verb{60 minutes}, \verb{60 min}}{Sixty minutes frequency, "60m"}
\item{90minutes, \verb{90 minutes}, \verb{90 min}}{Ninety minutes frequency, "90m"}
}
}
}
\usage{
time_intervals
}
\description{
A vector of common time series frequencies used in financial and economic data.
}
\keyword{datasets}
